fix(vscode): bundling `@hpcc-js/wasm` is necessary for the extension to work properly
