--- packages/vscode/tsconfig.src.json
@@ -1,34 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "declarationDir": "dist",
-    "tsBuildInfoFile": ".tsbuildinfo/src.tsbuildinfo",
-    "lib": [
-      "esnext",
-      "dom",
-      "WebWorker"
-    ]
-  },
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "**/node_modules",
-  ],
-  "references": [
-    {
-      "path": "../core"
-    },
-    {
-      "path": "../generators"
-    },
-    {
-      "path": "../language-server"
-    },
-    {
-      "path": "../layouts"
-    }
-  ]
-}
