--- packages/vscode/turbo.json
@@ -1,19 +1,0 @@
-{
-  "$schema": "https://turbo.build/schema.json",
-  "extends": ["//"],
-  "pipeline": {
-    "typecheck": {
-      "dependsOn": ["^build"],
-      "cache": false,
-      "outputs": []
-    },
-    "copy": {
-      "dependsOn": ["@likec4/vscode-preview#build"],
-      "outputs": ["dist/preview/**"]
-    },
-    "build": {
-      "dependsOn": ["copy", "^build"],
-      "outputs": ["dist/**", "!dist/preview/**"]
-    }
-  }
-}
